添加欢迎语

+++ include/library/Library.h
@@ -37,7 +37,7 @@ public:
     [[nodiscard]] const std::vector<Book>& getAllBooks() const;
 
     [[nodiscard]] std::vector<Book> searchBooksByISBN(std::string_view keyword) const;
-    std::vector<Book> getBooksBorrowedByUser(int userId);
+    std::vector<Book> getBooksBorrowedByUser(long long userId);
 
     // 图书搜索
     [[nodiscard]] std::vector<Book> searchBooksByTitle(std::string_view keyword) const;
@@ -46,7 +46,7 @@ public:
 
     bool borrowBook(long long userId, const QString& ISBN);
     bool returnBook(long long userId, const QString &ISBN);
-    bool isBookBorrowedByUser(int userId, const std::string& isbn);
+    bool isBookBorrowedByUser(long long userId, const std::string &isbn);
     // 数据持久化
     void saveToFile(const std::filesystem::path &userFile, const std::filesystem::path &bookFile) const;
     void loadFromFile(const std::filesystem::path &userFile, const std::filesystem::path &bookFile);
